requirements (1).txt
python-telegram-bot==20.3
requests==2.31.0
aiohttp==3.9.5
python-dotenv==1.0.1
